Move the clang compilation tests to gcc
They are really testing the gcc module behavior.
